+++ packages/terra-site/src/examples/card/CardRaised.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Card from 'terra-card';
+
+
+const CardDefault = () => (
+  <div>
+    <Card variant="raised">Hello World!!</Card>
+  </div>
+);
+
+export default CardDefault;